Make sized IRs use Sized constructor
Instead of layered SExprs. This trades some type safety for simplicity
of implementation. In particular those IRs can now be Monads, which
means we don't need to use Bound.Simple scopes leading to more code
reuse. It also means that our builtins don't need as many unknown sizes.

The downside of this approach is that there's are implicit invariants
for when sizes are necessary that need to hold.