Simplify: Treat current group as non-terminating
When eta-reducing e.g. the term

```
\x. f x
```

We need to check that evaluating `f` (_without calling it_) terminates.
Treating all globals as terminating should be the right thing to do, but
then leads e.g. to eta-reducing terms like

```
loop x = loop x
```

which isn't what we want. So we treat the currently processed group's
globals as non-terminating to avoid that but still get good
simplification.

With this in place it looks like the simplification works pretty well so
we probably don't need the planned hack of e.g. keeping track of all the
user-written lambdas (vs. generated by the compiler).