Implement some builtins using extern blocks
And make some changes necessary for that to work:

* Add language to primcalls to control the calling convention
* Make the normaliser not normalise some of the builtin operations that
it knows about to their extern representations since those are (more or
less) opaque to it.
* Split builtins in two files to be able to process builtins (int
operations) generated by the compiler before they are generated.

A problem with this is that extern functions don't seem to be inlined
even though we link the LLVM bytecode to the Sixten bytecode and
optimise afterwards (akin to link-time optimisation). This needs to be
investigated.