Add alignment type type representation
This means that the memory representation of e.g. `Tuple Byte (Ptr A)`
gets padding in the middle to make the pointer aligned in memory.

Type parameters are no longer merely ints representing the type's size
in memory, but they now contain both size and alignment information
(currently e.g. 62 bits of size and 2 bits of alignment packed into a
64-bit integer (on 64-bit targets)).  Product and sum types have to take
this into account (and not just add or max the sizes).