Normalise: Lazily evaluate instantiated variables
This shares the result of the normalisation of an instantiated variable,
such as in a lambda application, by using pre-solved existential
metavariables.

The change in SLam is because it previously didn't account for there
being solved metavariables in the syntax tree, but they can now be
introduced by the normaliser.