Syntax: Introduce implicit params with forall or @
Instead of curly braces.

This means that most polymorphic code can just use plain forall without
any curly braces, i.e. `forall v. ...` instead of the previous `forall
{v}. ...`.

Implicit applications are now done with `@` as in `f @arg` which frees
up curly braces in the syntax so they can be used for e.g. records.

It's also possible to use `@MyType -> ...` as a function arrow to match
the application syntax instead of forall.