Don't generalise non-constraints from def bodies
In generalisation we check what unsolved meta-variables are in both the
definition body and its type. We now only generalise the meta-variables
found in the type to implicit variables and leave the meta-variables
found in the body ungeneralised, unless they're constraints.

Generalising all variables found in bodies leads to surprising and weird
results sometimes, e.g. when generalising generated metavariables used
as implicit args of unit type.