Add local instance unification test
There's a problem with the elaborator: it doesn't abstract variables in
the expression and type at the same time, meaning that unified variables
in that appear both in the expression and the type don't get abstracted
in the type.

This would probably be fixed if elaboration is more baked into the
type-checker.