Add and use FreeVar types instead of MetaVars
... in the compiler passes that don't need the full generality of
MetaVars (unification etc), but just need unique variables with an
attached data or type.

We introduce two kinds of free variables representations:
FreeVar.FreeVar for when only some data needs to be attached, and
TypedFreeVar.FreeVar for when a type filled with free vars needs to be
attached. We change the lambda lifter, closure conversion, and return
direction analysis pass to use these new kinds of free variable.

This fixes #11.