Follow alias definitions before generating code
Aliases exist to make definitions like `funAlias = someFun` as efficient
as writing `someFun` (and not have it be closure-converted). This
required handling in the code generator, which wasn't done properly
before.

Before generating code, we now replace all globals that are aliases for
another global with the aliasee.

This means that AliasSigs don't need to be handled in the code
generator, so has a simplifying effect.