Shallowly skolemise when typechecking
The old prenex conversion/deep skolemisation function didn't actually do
deep skolemisation due to a bug, but just normal, shallow skolemisation,
and when I went on to fix it it turns out that it makes some tests fail.

So this commit fixes the deep skolemisation function, but also switches
to using an ordinary skolemisation function. While I was at it I made
the interface to the two functions the same so they can be more easily
interchanged for testing.