Make extern splices callbacks
Previously the following block would print both 10 and 20:

```
(C| if (true) { $(printInt 10); } else { $(printInt 20); } |)
```

The reason for this was that extern blocks would take the result of
running their splices as argument. This means that the splices were run
in advance.

We now instead make callbacks to Sixten at the use-site of the splices,
meaning that the above only prints 10, which is probably more sensible.

This doesn't handle splices of dynamic size, but that can come later. We
probably need to introduce some kind of macro syntax for out parameters
in that case.

Fixes #28.