Switch to parsix for parsing
And along the way switch to prettyprinter for pretty-printing and clean
up the error handling machinery a bit.

Parsix is a brand new parsing library written to overcome some
limitations of trifecta.

This commit doesn't actually do much but switch out the libraries,
but it opens up for a bright new future where:

* We can have parse error recovery since Parsix supports this
* We commit to using the text library (it's not perfect, but it feels
  better than the previous mixed approach) both when parsing and when
  generating error messages
* We get syntax highlighted error messages (and could emit that info
  for use in editor integration).