SLambda: Use explicit Anno type
The code generator expects there to be type annotations at certain
points in the expression tree, e.g. that all function and constructor
arguments start with a type annotation, but this was previously not
enforced by types.

This changes that by using an explicit `Anno` type where that is
expected, which gets rid of some `error`s and uses of "unknownSize" in
the backend.

There are still cases where we build invalid ASTs (e.g. with
"unknownSize") and rely on details of the code generator for it to work
(e.g. for the builtin pap and apply functions), but this at least
improves the situation.

Along the way we make some cleanups to the lambda lifter.