Fix generalisation issue #114
The issue was that when generalising a mixed signature-less and
signature-full binding group, the signatures would first be inferred but
not generalised, and then the signature-less definitions would be
inferred and generalised, and afterward everything generalised again.
But if a checked type-signature contains a metavariable and it's brought
into a signature-less definition during typechecking, this metavariable
will be generalised twice: first when generalising the signature-less
definitions, and then when generalising everything.

So to fix this, we make sure that the generalisation of the
signature-less definitions is not allowed to generalise any
metavariables created during the typechecking of the signatures, by
parameterising the generalisation functions by a metavariable predicate.

Fixes #114.