Pass allowed vars to pruning
Instead of abstracting before doing that. This fixes a problem where
pruning would fail when a disallowed variable's types mentioned an
allowed type, because the types of free variables don't partake in
abstraction, which means that the allowed free variable escapes when
abstracting the allowed variables, but not the disallowed ones (and not
in order).

The general rule is that abstraction has to be done in the same order as
instantiation if there are dependencies between the variables.