Perform cycle-detection on local lets
This check was removed a while ago, since it interfered with some
changes done to the typechecker.

It's now done as a pass immediately _after_ typechecking, which means it
doesn't have to deal with unsolved meta variables.

If there are cyclic definitions they will now be replaced with a failure
expression, which means that we won't violate assumptions further on in
the backend so we can gracefully keep going.

Fixes #117.