Implement a query-based compiler driver
This refactors the compiler from a push-based, more traditional,
pipeline, to a pull- and query-based architecture built on a build
system library made for the purpose, inspired by Shake and the Build
systems a la carte paper: https://github.com/ollef/rock

This will enable parallel and incremental builds in the future and will
help solve #10, #13, and #26.

This also implements some improvements that were necessary along the way
to get the tests to pass, for example:

* Recover from all errors (barring bugs), fixing #10.
* Improve logging with logging prefixes for selective printing.
* Fix some tests that were wrong.

There's still some stuff to refactor and fix up, but since the tests now
pass I think it's a good time to get it merged.