Remove support for alias signatures and definitions
Aliases were added such that definitions of the form `f = g` could be
compiled efficiently. The alias would make us see through the `f`,
noticing that it can be called as efficiently as `g`, instead of
creating a closure.

I think this functionality is better implemented by doing inlining as an
early simplification pass.